+++ src/com/company/circle.java
@@ -2,12 +2,12 @@ package com.company;
 import java.lang.Math;
 
 public class circle extends Shape {
-    double radius = this.size;
+    float radius = this.size;
 
     public boolean overlap(double dist, double size){
         return dist < size + this.size;
     }
-    public void set(double x1, double y1, double size1){
+    public void set(float x1, float y1, float size1){
         this.x = x1;
         this.y = y1;
         this.radius = size1;
